Gemma/requirements.txt
transformers
python-dotenv
notebook
streamlit
langchain